Improve Windows agent install and release packaging

+++ internal/disk/disk_unix.go
@@ -9,5 +9,5 @@ func FreeBytes(path string) (uint64, error) {
 	if err := syscall.Statfs(path, &stat); err != nil {
 		return 0, err
 	}
-	return stat.Bavail * uint64(stat.Bsize), nil
+	return uint64(stat.Bavail) * uint64(stat.Bsize), nil
 }
